backend/mineral_dna/requirements.txt
requests
aws-lambda-powertools
black
psycopg2-binary
boto3
